# packages/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the Diplom project",
        "title": "Diplom Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/api",
    "paths": {
...
        "/researchers": {
            "get": {
                "description": "Get a list of all researchers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researchers"
                ],
                "summary": "Get all researchers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Researcher"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new researcher",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researchers"
                ],
                "summary": "Create a new researcher",
                "parameters": [
                    {
                        "description": "Researcher object",
                        "name": "researcher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/researchers/{id}": {
            "get": {
                "description": "Get a single researcher by their ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researchers"
                ],
                "summary": "Get a researcher by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Researcher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "400": {
                        "description": "Invalid researcher ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Researcher not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing researcher's information",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researchers"
                ],
                "summary": "Update a researcher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Researcher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Researcher object",
                        "name": "researcher",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Researcher"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a researcher by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "researchers"
                ],
                "summary": "Delete a researcher",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Researcher ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/scholar/scrape": {
            "post": {
                "description": "Fetches all publications from a Google Scholar researcher profile URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scholar"
                ],
                "summary": "Scrape publications from Google Scholar",
                "parameters": [
                    {
                        "description": "Google Scholar URL with optional sort_by parameter ('pubdate', 'citations', or empty for default)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.ScholarScraperRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ScholarScraperResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.Publication": {
            "type": "object",
            "properties": {
                "abstract": {
                    "type": "string"
                },
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "citations": {
                    "type": "integer"
                },
                "issue": {
                    "type": "string"
                },
                "journal": {
                    "type": "string"
                },
                "pages": {
                    "type": "string"
                },
                "publisher": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "volume": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "handlers.ScholarScraperRequest": {
            "type": "object",
            "properties": {
                "sort_by": {
                    "description": "Optional: \"pubdate\", \"citations\", \"title\" or empty for default",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.ScholarScraperResponse": {
            "type": "object",
            "properties": {
                "publications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.Publication"
                    }
                },
                "sort_by": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        },
...
        "models.LocalizedString": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "ru": {
                    "type": "string"
                }
            }
        },
...
        "models.Publication": {
            "type": "object",
            "properties": {
                "authors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "journal": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "models.Researcher": {
            "type": "object",
            "properties": {
                "bio": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "name": {
                    "$ref": "#/definitions/models.LocalizedString"
                },
                "photo": {
                    "type": "string"
                },
                "profiles": {
                    "$ref": "#/definitions/models.ResearcherProfiles"
                },
                "publications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Publication"
                    }
                }
            }
        },
        "models.ResearcherProfiles": {
            "type": "object",
            "properties": {
                "googleScholar": {
                    "type": "string"
                },
                "orcid": {
                    "type": "string"
                },
                "publons": {
                    "type": "string"
                },
                "researchgate": {
                    "type": "string"
                },
                "scopus": {
                    "type": "string"
                }
            }
        },
...
    }
}